requirements.txt
Django==1.8.5
psycopg2==2.6.1
requests==2.8.1
pytz==2015.7
django-versatileimagefield==1.5
djangorestframework==3.3.3
